% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flood_val.R
\name{find_NWS}
\alias{find_NWS}
\title{Get National Weather Service (NWS) flood stage/discharge levels for gages.}
\usage{
find_NWS(site_no, type)
}
\arguments{
\item{site_no}{Character vector with USGS gage IDs of stream gage sites to
pull.}

\item{type}{Character string with the type of flood stage to be used. Can be
one of four options: "action", "flood", "moderate", and "major".}
}
\value{
Data frame of gage IDs and the corresponding NWS flood value, if
available.
}
\description{
Use National Weather Service designated flood stages/discharges as flood
thresholds. These come in four levels: "action", "flood", "moderate", and
"major". Note that most USGS gages do not have these values specified (or may
not have all levels) so using this definition of the flood threshold can
severely limit the sample size of the data output.
}
\examples{

va_counties <- get_county_cd("Virginia")
va_gages <- get_gages(va_counties, start_date = "2015-01-01",
                      end_date = "2015-12-31")
va_NWS <- find_NWS(site_no = va_gages$site_no, type = "moderate")

}

