% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flood_val.R
\name{find_Q2}
\alias{find_Q2}
\title{Get median flood for each gage}
\usage{
find_Q2(site_no)
}
\arguments{
\item{site_no}{Character vector with USGS gage IDs of stream gage sites to
pull.}
}
\value{
A data frame with median flood values (Q2) and the number of years of
  data used to compute this value.
}
\description{
This function will get annual maximum flow series for each USGS gage and
compute median flood (Q2) to serve as flood threshold
}
\examples{

miami_gages <- gage_extract("12086", start_date = "2000-01-01",
                            end_date = "2009-12-31")
miami_Q2 <- find_Q2(site_no = miami_gages$site_no)

}

