% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_county_cd}
\alias{get_county_cd}
\title{Get all FIPS county codes within states}
\usage{
get_county_cd(state)
}
\arguments{
\item{state}{Character vector giving the name of state or states for which
you would like to get county FIPS codes.}
}
\value{
A character vector with the 5-digit FIPS codes for all counties
   within the specified state or states.
}
\description{
This function will return all county FIPS codes for all counties within a
state or states.
}
\examples{
 get_county_cd("Virginia")
 get_county_cd(c("North Carolina", "South Carolina"))

}

