% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_results.R
\name{map_flood}
\alias{map_flood}
\title{Function maps data, either by gage or by county}
\usage{
map_flood(flood_stats)
}
\arguments{
\item{flood_stats}{Data frame of flood analysis results, output of
\code{run_flood} function.}
}
\value{
A map of the state(s) analyzed showing counties and gages color coded
  based on flood magnitude, depending on the type of data in flood_stats.
}
\description{
Displays a state or multi-state map summarizing flood analysis results either
by gage or county.
}
\examples{
#Use Q2 as flood threshold and get get gage-level output
va_floods <- run_flood(state = "Virginia", start_date = "2015-01-01",
                      end_date = "2015-12-31", threshold = "Q2", output = "gage")
#Map results by gage
map_flood(va_floods)


#Use NWS flood thresholds and get county-level output
va_floods <- run_flood(state = "Virginia", start_date = "2015-01-01",
                      end_date = "2015-12-31", threshold = "NWS", flood_type = "action",
                      output = "county")
#Map results by county
map_flood(va_floods)

}

