% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_results.R
\name{map_gage}
\alias{map_gage}
\title{Maps flood data by gage}
\usage{
map_gage(flood_stats, date = "")
}
\arguments{
\item{flood_stats}{Data frame of flood analysis results, by gage.}
}
\value{
A map of USGS gages color coded by maximum flood magnitude.
}
\description{
Creates a state level map of flood analysis output by USGS gage. Gages are
color coded based on maximum flood magnitude (for flood threshold = "Q2").
If the flood threshold is "NWS", points are binary coded based on flood
occurrence (e.g. yes/no).
}
