% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flood_val.R
\name{construct_prob_plot}
\alias{construct_prob_plot}
\title{Construct probability plot using the Weibull plotting method}
\usage{
construct_prob_plot(vals)
}
\arguments{
\item{vals}{A numeric vector with peak discharge values ... [more on what
this column of the data from readNWISpeak gives]}
}
\description{
Construct probability plot using the Weibull plotting method
}

