% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_data.R
\name{flood_analysis}
\alias{flood_analysis}
\title{Compute flood statistics}
\usage{
flood_analysis(flow_data, peaks, gages)
}
\description{
Takes flow data and computes flood statistics based on selected flood
definition.
}

