% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_gages}
\alias{get_gages}
\title{Get all gage site numbers for a county}
\usage{
get_gages(county_cd, start_date, end_date)
}
\arguments{
\item{county_cd}{Character vector with the county FIPS code}

\item{start_date}{Character string with the starting date, using "YYYY-MM-DD"
notation.}

\item{end_date}{Character string with the end date, using "YYYY-MM-DD"
notation.}
}
\value{
A dataframe that gives the following variables for stream gages within
   the county and time range:
   \itemize{
   \item{agency_cd: }
   \item{site_no: }
   \item{station_nm: }
   \item{site_tp_cd: }
   \item{dec_lat_va: }
   \item{dec_long_va: }
   \item{county_cd: The five-digit FIPS code of the county in which the gage is located}
   }
}
\description{
Pulls gage numbers of all gages with discharge data within a county and
within the specified date range.
}
\examples{

get_gages("12086", start_date = "1988-01-01", end_date = "2015-01-01")

va_counties <- get_county_cd("Virginia")
va_gages <- get_gages(va_counties, start_date = "2015-01-01",
                      end_date = "2015-12-31")
# Equivalent with piping
va_gages <- get_county_cd("Virginia") \%>\%
   get_gages(start_date = "2015-01-01", end_date = "2015-12-31")

}
